+++ src/main/java/org/truffleruby/core/queue/QueueNodes.java
@@ -9,13 +9,20 @@
  */
 package org.truffleruby.core.queue;
 
+import com.oracle.truffle.api.dsl.ImportStatic;
+import com.oracle.truffle.api.library.CachedLibrary;
 import org.truffleruby.annotations.CoreMethod;
 import org.truffleruby.builtins.CoreMethodArrayArgumentsNode;
 import org.truffleruby.builtins.CoreMethodNode;
 import org.truffleruby.annotations.CoreModule;
 import org.truffleruby.builtins.NonStandard;
+import org.truffleruby.core.array.ArrayGuards;
+import org.truffleruby.core.array.RubyArray;
+import org.truffleruby.core.array.library.ArrayStoreLibrary;
 import org.truffleruby.core.cast.BooleanCastWithDefaultNode;
+import org.truffleruby.core.cast.ToANode;
 import org.truffleruby.core.klass.RubyClass;
+import org.truffleruby.language.NotProvided;
 import org.truffleruby.language.RubyBaseNodeWithExecute;
 import org.truffleruby.language.RubyNode;
 import org.truffleruby.annotations.Visibility;
@@ -187,6 +194,27 @@ public abstract class QueueNodes {
             return self;
         }
 
+    }
+
+    @CoreMethod(names = "initialize", visibility = Visibility.PRIVATE, optional = 1)
+    @ImportStatic(ArrayGuards.class)
+    public abstract static class InitializeNode extends CoreMethodArrayArgumentsNode {
+
+        @Specialization
+        protected RubyQueue execute(RubyQueue self, NotProvided enumerable) {
+            return self;
+        }
+
+        @Specialization(guards = "wasProvided(enumerable)")
+        protected RubyQueue execute(RubyQueue self, Object enumerable,
+                @CachedLibrary(limit = "storageStrategyLimit()") ArrayStoreLibrary stores,
+                @Cached ToANode toANode) {
+            final RubyArray rubyArray = toANode.executeToA(enumerable);
+            final Object[] array = stores.boxedCopyOfRange(rubyArray.getStore(), 0, rubyArray.size);
+            self.queue.addAll(array);
+
+            return self;
+        }
     }
 
     @CoreMethod(names = "marshal_dump")