+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Compatibility:
 * Fix desctructuring of a single block argument that implements `#to_ary` dynamically (#2719, @andrykonchin).
 * Fix `Kernel#Complex` and raise exception when an argument is formatted incorrectly (#2765, @andrykonchin).
 * Add `#public?`, `#private?` and `#protected?` methods for `Method` and `UnboundMethod` classes (@andrykonchin).
+* Add optional argument to `Thread::Queue.new` (@andrykonchin).
 
 Performance: